requirements.txt
aiohttp==3.6.2
aiohttp-security==0.4.0
aiohttp-session==2.9.0
aiopg==1.0.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
multidict==4.7.6
passlib==1.7.2
pip-autoremove==0.9.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
SQLAlchemy==1.3.17
yarl==1.4.2
